+++ src/backend/resolvers/mutations/relationalResolvers/obmAndPayer/delete.js
@@ -1,0 +1,28 @@
+const { ObjectId } = require('mongodb')
+
+const deleteObmAndPayerConnection = async (
+  parent,
+  { input: { _id } },
+  { pulseCoreDb, pulseDevDb, mongoClient }
+) => {
+  _id = ObjectId(_id)
+
+  const session = mongoClient.startSession()
+
+  let deletedConnection
+
+  await session.withTransaction(async () => {
+    deletedConnection = await pulseCoreDb
+      .collection('JOIN_obms_payers')
+      .findOneAndDelete({ _id }, { session })
+      .then(({ value }) => value)
+
+    await pulseDevDb
+      .collection('NEW_obmsPayers')
+      .deleteOne({ _id }, { session })
+  })
+
+  return deletedConnection
+}
+
+module.exports = deleteObmAndPayerConnection