+++ src/backend/typeDefs/mutations/relationalTypeDefs/obmAndPayer/delete.js
@@ -1,0 +1,9 @@
+const { gql } = require('apollo-server-express')
+
+const deletePathwaysAndPersonConnectionTypeDefs = gql`
+  input DeleteObmAndPayerConnectionInput {
+    _id: ID!
+  }
+`
+
+module.exports = deletePathwaysAndPersonConnectionTypeDefs