+++ src/frontend/Orion/shared/AccountModals/AccountModalButton/AccountModalContent/ConnectionsSection/utils/state-abbreviations.js
@@ -49,4 +49,6 @@ export default [
   'WV',
   'WI',
   'WY',
+  'PR',
+  'DC',
 ]