+++ src/frontend/components/BusinessObjectModal/OncologyBenefitManagerModal/relational-widgets/ObmPayersWidget/ObmPayersWidget.js
@@ -1,68 +1,95 @@
 import React, { useEffect, useState } from 'react'
 import { useQuery } from '@apollo/react-hooks'
 import _ from 'lodash'
+import styled from '@emotion/styled'
 
+import useObmAndPayerConnections from 'frontend/hooks/useObmPayerConnections'
+import manualBoModalLockOverlay from 'frontend/components/BusinessObjectModal/shared/widget/manualBoModalLockOverlay'
+
+import ConnectionPanel from './ConnectionPanel'
 import ConnectionsList from './ConnectionsList'
-import ConnectionForm from './ConnectionForm'
 
-import {
-  GET_PAYER_ORGANIZATIONS,
-  GET_JOIN_OBMS_AND_PAYERS,
-  GET_BOOKS,
-} from '../../../../../api/queries'
+import { GET_PAYER_ORGANIZATIONS, GET_BOOKS } from '../../../../../api/queries'
+
+const WidgetContainer = styled.div({
+  display: 'flex',
+  width: '100%',
+})
+
+const WIDGET_TITLE = 'Payer Connections'
 
 const ObmPayersWidget = ({ entity }) => {
-  const [stagedConnections, stageConnections] = useState([])
   const [selectedConnectionId, selectConnectionId] = useState(null)
+  const [
+    isNewConnectionBeingCreated,
+    setWhetherNewConnectionBeingCreated,
+  ] = useState(false)
+  const [anyUnsavedChanges, setWhetherUnsavedChanges] = useState(false)
 
   const { data: payersData, loading: payersLoading } = useQuery(
     GET_PAYER_ORGANIZATIONS
   )
   const { data: booksData, loading: booksLoading } = useQuery(GET_BOOKS)
-  const { data: connectionsData, loading: connectionsLoading } = useQuery(
-    GET_JOIN_OBMS_AND_PAYERS,
-    {
-      variables: { obmId: entity._id },
-    }
-  )
+  const {
+    data: connections,
+    loading: connectionsLoading,
+  } = useObmAndPayerConnections({ obmId: entity._id })
 
   useEffect(() => {
-    if (!payersLoading && !connectionsLoading && !booksLoading) {
-      const connections = Object.values(connectionsData)[0]
-      const { _id: firstConnectionId } = connections[0]
-
+    if (!payersLoading && !booksLoading && !connectionsLoading) {
+      const { _id: firstConnectionId } = connections[0] || {}
       selectConnectionId(firstConnectionId)
-      stageConnections(connections)
     }
   }, [payersLoading, connectionsLoading, booksLoading])
 
   if (payersLoading || connectionsLoading || booksLoading) return 'Loading...'
 
-  let allPayers = []
-  let payerOrgById = {}
-  if (!payersLoading) {
-    allPayers = Object.values(payersData)[0]
-
-    payerOrgById = _.mapValues(_.keyBy(allPayers, '_id'), 'organization')
+  const createConnectionHandler = () => {
+    if (anyUnsavedChanges) {
+      alert(
+        "You have unsaved changes! Please save or cancel the connection you're on."
+      )
+    } else {
+      setWhetherNewConnectionBeingCreated(true)
+    }
   }
 
+  let allPayers = []
+  if (!payersLoading) allPayers = Object.values(payersData)[0]
+  const payerOrgById = _.mapValues(_.keyBy(allPayers, '_id'), 'organization')
+
+  let allBooks = []
+  if (!booksLoading) allBooks = Object.values(booksData)[0]
+
+  manualBoModalLockOverlay(anyUnsavedChanges)
+
   return (
-    <div style={{ flex: '1 0 auto', display: 'flex' }}>
+    <WidgetContainer>
       <ConnectionsList
+        connections={connections}
+        widgetTitle={WIDGET_TITLE}
+        createConnectionHandler={createConnectionHandler}
         selectedConnectionId={selectedConnectionId}
         selectConnectionId={selectConnectionId}
-        stagedConnections={stagedConnections}
-        stageConnection={stageConnections}
+        isNewConnectionBeingCreated={isNewConnectionBeingCreated}
+        anyUnsavedChanges={anyUnsavedChanges}
         payerOrgById={payerOrgById}
       />
 
-      <ConnectionForm
-        connection={stagedConnections.find(
-          ({ _id }) => _id === selectedConnectionId
-        )}
-        allPayers={allPayers}
+      <ConnectionPanel
+        connection={connections.find(({ _id }) => _id === selectedConnectionId)}
+        connections={connections}
+        payerOrgById={payerOrgById}
+        isNewConnectionBeingCreated={isNewConnectionBeingCreated}
+        obmId={entity._id}
+        setWhetherNewConnectionBeingCreated={
+          setWhetherNewConnectionBeingCreated
+        }
+        setWhetherUnsavedChanges={setWhetherUnsavedChanges}
+        selectConnectionId={selectConnectionId}
+        allBooks={allBooks}
       />
-    </div>
+    </WidgetContainer>
   )
 }